--- BookReview/Utils + Extensions/Extensions/CharacterSetExtensions.swift
@@ -1,21 +1,0 @@
-//
-//  CharacterSetExtensions.swift
-//  BookReview
-//
-//  Created by Felipe Arturo. Perez Camargo on 20/12/20.
-//  Copyright © 2020 Televisa. All rights reserved.
-//
-
-import Foundation
-
-extension CharacterSet {
-    static let urlQueryValueAllowed: CharacterSet = {
-        let generalDelimitersToEncode = ":#[]@" // does not include "?" or "/" due to RFC 3986 - Section 3.4
-        let subDelimitersToEncode = "!$'()*+,;"// does not include "="  or "&" due API normaly has troubles with them
-
-        var allowed = CharacterSet.urlQueryAllowed
-        allowed.remove(charactersIn: generalDelimitersToEncode + subDelimitersToEncode)
-
-        return allowed
-    }()
-}
