+++ BookReview/Utils + Extensions/Utils/CellIdentifiers.swift
@@ -8,6 +8,8 @@
 
 import Foundation
 
+// To keep this string organized for the basic cell provided by system
+// Custom cells should contain a static property called "identifier"
 struct CellIdentifiers {
     static let cellIdentifier = "basicCell"
 }
